\name{FamilyRank-package}
\alias{FamilyRank-package}
\alias{FamilyRank}
\docType{package}
\title{
\packageTitle{FamilyRank}
}
\description{
\packageDescription{FamilyRank}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{FamilyRank}
\packageIndices{FamilyRank}
The main function is \code{\link{familyRank}}.
}
\author{
\packageAuthor{FamilyRank}

Maintainer: \packageMaintainer{FamilyRank}
}
\references{ADD REFERENCE}
\keyword{ package }
